apps/elma/telegram_bot/requirements.txt
python-telegram-bot===13.9
python-dotenv===0.19.2
peewee===3.14.8
pymysql===1.0.2
Faker===10.0.0
pika===1.2.0
